Monkey/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "monkey-collection"
version = "0.1.0"
description = "GUI World Modeling data collection pipeline"
requires-python = ">=3.10"
dependencies = [
    "pyyaml>=6.0",
    "Pillow>=10.0",
    "loguru>=0.7",
]

[project.optional-dependencies]
annotation = [
    "openai>=1.0",
]

[tool.setuptools.packages.find]
include = ["collection*"]

[project.scripts]
monkey-collect = "collection.cli:main"
